requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
flask>=3.0.0
